Error fix in right_rotate.
Functions right_rotate and left_rotate should be symmetrical. So that we can replace left with right and vice versa and get the second function. This is not the case, so I think this is a mistake. I also tested it myself and after the fix, my code started working correctly.